+++ changes.md
@@ -3,7 +3,10 @@
 ## Main wip
 ## 2.0.0
 * Refactored Ramp to be an orchestrator
-* Added BaseCoffin 
+* Added BaseCoffin
+* Added ContainerDoor
+* Added ContainerRamp wich uses container door and RampGreebled
+* Added Container Greeble Hack rough example container_greeble_hack.py
 
 ## 1.2.0
 * Added FloorTile